+++ src/components/project/billing.tsx
@@ -1,4 +1,8 @@
 import React, { useState } from "react";
+import { useLocation } from "react-router-dom";
+
+import Path from "./path";
+
 import { TfiBarChart } from "react-icons/tfi";
 import { MdOutlineElectricalServices, MdOutlineHistory } from "react-icons/md";
 import { HiOutlineWallet } from "react-icons/hi2";
@@ -6,8 +10,10 @@ import { AiTwotoneThunderbolt } from "react-icons/ai";
 import { LiaCalendarCheck } from "react-icons/lia";
 
 export default function Billing() {
-  const [activeMenu, setActiveMenu] = useState(null);
+  const [activeMenu, setActiveMenu] = useState("plan");
   const [isPopupOpen, setIsPopupOpen] = useState(false);
+  const location = useLocation();
+  const pathName = location.pathname;
   const handleMenuClick = (key) => {
     setActiveMenu(key);
   };
@@ -18,27 +24,63 @@ export default function Billing() {
     setIsPopupOpen(false);
   };
   const menuItems = [
-    { key: "usage", label: "Usage", icon: <TfiBarChart /> },
-    { key: "plan", label: "Plan", icon: <MdOutlineElectricalServices /> },
-    { key: "add-ons", label: "Add-ons", icon: <MdOutlineElectricalServices /> },
-    { key: "history", label: "History", icon: <MdOutlineHistory /> },
-    { key: "payment", label: "Payment methods", icon: <HiOutlineWallet /> },
+    {
+      key: "usage",
+      label: "Usage",
+      icon: <TfiBarChart />,
+    },
+    {
+      key: "plan",
+      label: "Plan",
+      icon: <MdOutlineElectricalServices />,
+    },
+    {
+      key: "add-ons",
+      label: "Add-ons",
+      icon: <MdOutlineElectricalServices />,
+    },
+    {
+      key: "history",
+      label: "History",
+      icon: <MdOutlineHistory />,
+    },
+    {
+      key: "payment",
+      label: "Payment methods",
+      icon: <HiOutlineWallet />,
+    },
   ];
+
   const invoiceData = [
-    { name: "Discover Plan", amount: "$0.00" },
-    { name: "Overages", amount: "$0.00" },
-    { name: "Multi-region Transaction Broadcast", amount: "$0.00" },
-    { name: "Multi-region Transaction Broadcast", amount: "$0.00" },
-    { name: "Total", amount: "$0.00" },
+    {
+      name: "Discover Plan",
+      amount: "$0.00",
+    },
+    {
+      name: "Overages",
+      amount: "$0.00",
+    },
+    {
+      name: "Multi-region Transaction Broadcast",
+      amount: "$0.00",
+    },
+    {
+      name: "Multi-region Transaction Broadcast",
+      amount: "$0.00",
+    },
+    {
+      name: "Total",
+      amount: "$0.00",
+    },
   ];
   function Plan() {
     return (
-      <div className=" text-black ">
+      <div id="billing" className=" text-black  grid pb-20">
         <div className=" flex justify-between mb-10">
           <div className="font-normal text-lg px-5 py-2">Plan : Discover</div>
           <div className=" flex">
             <button
-              className=" flex items-center border-black text-white bg-purple-600 rounded-lg"
+              className=" flex items-center mx-3 border-black text-white bg-purple-600 rounded-lg"
               onClick={openPopup}
             >
               <div className=" mx-3 flex items-center ">
@@ -64,9 +106,9 @@ export default function Billing() {
 
         <div className=" mt-5 mb-14">
           <div className=" mb-2 text-sm">Upcoming Invoice</div>
-          <div className=" flex justify-between border border-gray-200 rounded-lg py-10 px-8">
-            <div>
-              <div className="flex items-center text-sm font-semibold">
+          <div className=" flex gap-6 border border-gray-200 rounded-lg py-10 px-8">
+            <div className=" border-r-2 w-full">
+              <div className="flex items-center text-sm font-semibold ">
                 Billing Period
               </div>
               <div className="flex items-center">
@@ -74,8 +116,7 @@ export default function Billing() {
                 Aug 27, 2023 - Sep 27, 2023
               </div>
             </div>
-            <span className="text-gray-300  text-5xl">|</span>
-            <div className="">
+            <div className="w-full">
               <div>Total</div>
               <div>$0.00</div>
             </div>
@@ -102,6 +143,7 @@ export default function Billing() {
   }
   return (
     <div className=" bg-white rounded-lg shadow-md p-6">
+      <Path pathname={location.pathname} />
       <h2 className=" text-2xl font-semibold mb-4 text-black">
         Usage & Billing
       </h2>
@@ -113,9 +155,9 @@ export default function Billing() {
                 onClick={() => handleMenuClick(item.key)}
                 className={`${
                   activeMenu === item.key
-                    ? "font-semibold border-b-2 border-blue-500"
-                    : ""
-                } focus:outline-none`}
+                    ? "border-b-2 border-blue-500"
+                    : "border-transparent"
+                } focus:outline-none hover:bg-gray-100 hover:rounded-lg`}
               >
                 <div className=" flex items-center">
                   {item.icon}
@@ -126,8 +168,8 @@ export default function Billing() {
           ))}
         </ul>
       </nav>
-      ;
-      <div className=" container mx-auto p-4">
+
+      <div className=" container  mx-auto p-5 value">
         {activeMenu === "usage" && <div>홈 페이지 컨텐츠</div>}
         {activeMenu === "plan" && <Plan></Plan>}
         {activeMenu === "add-ons" && <div>소개 페이지 컨텐츠</div>}
